+++ one_fm/one_fm/custom/shift_request.json
@@ -145,7 +145,7 @@
    "hide_border": 0,
    "hide_days": 0,
    "hide_seconds": 0,
-   "idx": 15,
+   "idx": 16,
    "ignore_user_permissions": 0,
    "ignore_xss_filter": 0,
    "in_global_search": 0,
@@ -205,7 +205,7 @@
    "hide_border": 0,
    "hide_days": 0,
    "hide_seconds": 0,
-   "idx": 16,
+   "idx": 17,
    "ignore_user_permissions": 0,
    "ignore_xss_filter": 0,
    "in_global_search": 0,
@@ -265,7 +265,7 @@
    "hide_border": 0,
    "hide_days": 0,
    "hide_seconds": 0,
-   "idx": 18,
+   "idx": 19,
    "ignore_user_permissions": 0,
    "ignore_xss_filter": 0,
    "in_global_search": 0,
@@ -325,7 +325,7 @@
    "hide_border": 0,
    "hide_days": 0,
    "hide_seconds": 0,
-   "idx": 17,
+   "idx": 18,
    "ignore_user_permissions": 0,
    "ignore_xss_filter": 0,
    "in_global_search": 0,
@@ -385,7 +385,7 @@
    "hide_border": 0,
    "hide_days": 0,
    "hide_seconds": 0,
-   "idx": 13,
+   "idx": 14,
    "ignore_user_permissions": 0,
    "ignore_xss_filter": 0,
    "in_global_search": 0,
@@ -421,9 +421,136 @@
    "width": null
   }
  ],
- "custom_perms": [],
+ "custom_perms": [
+  {
+   "_assign": null,
+   "_comments": null,
+   "_liked_by": null,
+   "_user_tags": null,
+   "amend": 0,
+   "cancel": 0,
+   "create": 1,
+   "creation": "2018-04-13 16:32:27.974273",
+   "delete": 0,
+   "docstatus": 0,
+   "email": 1,
+   "export": 1,
+   "idx": 3,
+   "if_owner": 0,
+   "import": 0,
+   "modified": "2022-06-21 08:29:12.248404",
+   "modified_by": "s.shaikh@armor-services.com",
+   "name": "88b8d0b4bb",
+   "owner": "Administrator",
+   "parent": "Shift Request",
+   "parentfield": "permissions",
+   "parenttype": "DocType",
+   "permlevel": 0,
+   "print": 1,
+   "read": 1,
+   "report": 1,
+   "role": "HR User",
+   "select": 0,
+   "set_user_permissions": 0,
+   "share": 1,
+   "submit": 1,
+   "write": 1
+  },
+  {
+   "_assign": null,
+   "_comments": null,
+   "_liked_by": null,
+   "_user_tags": null,
+   "amend": 1,
+   "cancel": 1,
+   "create": 1,
+   "creation": "2018-04-13 16:32:27.974273",
+   "delete": 1,
+   "docstatus": 0,
+   "email": 1,
+   "export": 1,
+   "idx": 2,
+   "if_owner": 0,
+   "import": 0,
+   "modified": "2022-06-21 08:29:12.325404",
+   "modified_by": "s.shaikh@armor-services.com",
+   "name": "95852459e8",
+   "owner": "Administrator",
+   "parent": "Shift Request",
+   "parentfield": "permissions",
+   "parenttype": "DocType",
+   "permlevel": 0,
+   "print": 1,
+   "read": 1,
+   "report": 1,
+   "role": "HR Manager",
+   "select": 0,
+   "set_user_permissions": 0,
+   "share": 1,
+   "submit": 1,
+   "write": 1
+  },
+  {
+   "_assign": null,
+   "_comments": null,
+   "_liked_by": null,
+   "_user_tags": null,
+   "amend": 0,
+   "cancel": 0,
+   "create": 1,
+   "creation": "2018-04-13 16:32:27.974273",
+   "delete": 0,
+   "docstatus": 0,
+   "email": 1,
+   "export": 1,
+   "idx": 1,
+   "if_owner": 0,
+   "import": 0,
+   "modified": "2022-06-21 08:29:12.331484",
+   "modified_by": "s.shaikh@armor-services.com",
+   "name": "2fbdd67c34",
+   "owner": "Administrator",
+   "parent": "Shift Request",
+   "parentfield": "permissions",
+   "parenttype": "DocType",
+   "permlevel": 0,
+   "print": 1,
+   "read": 1,
+   "report": 1,
+   "role": "Employee",
+   "select": 0,
+   "set_user_permissions": 0,
+   "share": 1,
+   "submit": 1,
+   "write": 1
+  }
+ ],
  "doctype": "Shift Request",
  "property_setters": [
+  {
+   "_assign": null,
+   "_comments": null,
+   "_liked_by": null,
+   "_user_tags": null,
+   "creation": "2022-08-03 14:06:20.338293",
+   "default_value": null,
+   "doc_type": "Shift Request",
+   "docstatus": 0,
+   "doctype_or_field": "DocField",
+   "field_name": "approver",
+   "idx": 0,
+   "modified": "2022-08-03 14:06:20.338293",
+   "modified_by": "Administrator",
+   "name": "Shift Request-approver-ignore_user_permissions",
+   "owner": "Administrator",
+   "parent": null,
+   "parentfield": null,
+   "parenttype": null,
+   "property": "ignore_user_permissions",
+   "property_type": "Check",
+   "row_name": null,
+   "value": "1"
+  },
   {
    "_assign": null,
    "_comments": null,
